Fix terminal in long rules bug in chomsky normal form
Add sections called 'remove unit rules' and 'remove terminal in
long rules' to chomsky_normal_form tree transform function. This
allows the transform to work when there is A -> B unit rule in
the original grammar, or when there is a terminal character in
long rule, such as A -> BcD in the origianl grammar. Also add
some regression tests.